+++ src/crud/update-status.js
@@ -1,5 +1,5 @@
 'use strict';
-var outputMap = require('../output-map');
+var outputMap = require('../output');
 var applyMaps = require('../swagger/router/step-maps');
 var getValidateFunction = require('./@shared/get-validate-function');
 var schemaName = 'update';
