+++ src/output/object-path-validation.js
